requirements.txt
numpy>=1.24.0
flwr[simulation]>=1.8.0
ray>=2.0.0
torch>=2.0.0
torchvision>=0.15.0
web3>=6.0.0
python-dotenv>=1.0.0
requests>=2.31.0